requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.23
celery==3.1.23
certifi==2016.2.28
cffi==1.5.2
cryptography==1.3.1
decorator==4.0.9
enum34==1.1.2
facebook-sdk==0.4.0
Flask==0.10.1
Flask-Cache==0.13.1
flask-canvas==0.1
Flask-Login==0.3.2
Flask-OpenID==1.2.5
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.0
foursquare==1!2015.5.26
httplib2==0.9.2
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
MarkupSafe==0.23
mimerender==0.6.0
MySQL-python==1.2.5
oauthlib==1.0.3
pbr==1.8.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
python-dateutil==2.5.1
python-instagram==1.3.2
python-mimeparse==1.5.1
python-openid==2.2.5
pytz==2016.3
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.0.12
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
Tempita==0.5.2
tornado==4.3
Werkzeug==0.11.4
wheel==0.24.0
yelpapi==1.3
